requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
albucore==0.0.19
albumentations==1.4.20
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blessed==1.20.0
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1707229174982/work/certifi
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
datasets==3.0.1
decord==0.6.0
deepspeed==0.14.4
diffusers==0.27.2
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
docker-pycreds==0.4.0
einops==0.8.0
email_validator==2.1.1
eval_type_backport==0.2.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.14.0
fonttools==4.52.4
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
gpustat==1.1.1
gradio==4.32.0
gradio_client==0.17.0
grpcio==1.65.4
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.2
idna @ file:///croot/idna_1714398848350/work
imageio==2.27.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
imutils==0.5.4
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lmdb==1.5.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
ninja==1.11.1.1
nltk==3.9.1
numpy==1.26.4
nvidia-ml-py==12.555.43
omegaconf==2.3.0
openai==1.51.2
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
orjson==3.10.3
packaging==24.0
pandas==2.2.2
peft==0.11.1
pillow @ file:///croot/pillow_1714398848491/work
platformdirs==4.2.2
propcache==0.2.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==17.0.0
pydantic==2.7.2
pydantic_core==2.18.3
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
pyre-extensions==0.0.23
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==0.7.3
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests @ file:///croot/requests_1716902831423/work
rich==13.7.1
rpds-py==0.18.1
ruff==0.4.6
safetensors==0.4.3
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.3.1
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
stringzilla==3.10.5
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
timm==0.6.12
tokenizers==0.19.1
tomlkit==0.12.0
toolz==0.12.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.66.4
transformers==4.41.2
triton==2.3.1
typer==0.12.3
typing-inspect==0.9.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.1
ujson==5.10.0
ultralytics==8.0.180
urllib3 @ file:///croot/urllib3_1715635851070/work
uvicorn==0.30.0
uvloop==0.19.0
wandb==0.17.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.3
xformers==0.0.16
xxhash==3.5.0
yarl==1.14.0
zipp==3.19.0
zstandard==0.23.0
